# student-management-service/swagger.yaml
openapi: 3.0.0
info:
  title: Student Management Service API
  version: 1.0.0
  description: API for managing student enrollments, payments, certificates, courses, quizzes, and more
paths:
...
  /quizzes/{id}:
    get:
      summary: Get quiz details
      tags:
        - Quizzes
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Update a quiz
      tags:
        - Quizzes
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete a quiz
      tags:
        - Quizzes
      security:
        - CognitoUserPools: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CognitoUserPools:
      type: apiKey
      name: Authorization
      in: header